pyproject.toml
[tool.poetry]
name = "app-drf-initial"
version = "0.1.0"
description = ""
authors = ["Felipe Savaris <fesavaris@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
